% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump-midi.R
\name{dump_midi}
\alias{dump_midi}
\title{Dump a MIDI file}
\usage{
dump_midi(filename, print = TRUE)
}
\arguments{
\item{filename}{The path to the MIDI file.}

\item{print}{If \code{TRUE}, print the contents of the MIDI file to the console.}
}
\value{
The contents of the MIDI file as a character vector is invisibly returned.
}
\description{
Dump a MIDI file
}
\keyword{internal}
