% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipopaplot-package.R
\docType{package}
\name{pipopaplot-package}
\alias{pipopaplot}
\alias{pipopaplot-package}
\title{pipopaplot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Other contributors:
\itemize{
  \item Craig Stuart Sapp [copyright holder]
}

}
\keyword{internal}
