% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{rollup}
\alias{rollup}
\title{Aggregate or transform note data before sonification}
\usage{
rollup(d, by, .fun = base::mean)
}
\arguments{
\item{d}{A data frame returned by \code{\link[=as_notes]{as_notes()}}.}

\item{by}{A variable or expression used for grouping within each
\verb{(channel, group)} pair.}

\item{.fun}{A summary function applied to numeric columns.
Defaults to \code{\link[base:mean]{base::mean()}}.}
}
\value{
A data frame containing the same key columns
(\code{x}, \code{y}, \code{channel}, \code{group}, \code{duration}, \code{velocity})
and suitable for use in \code{\link[=sonify]{sonify()}}.
}
\description{
\code{rollup()} provides a simple mechanism to summarize or reshape
note-level data prior to sonification.
By default, it performs numeric aggregation within each
\verb{(channel, group, by)} combination.
}
\details{
Although \code{rollup()} uses summarization by default,
any transformation that returns a data frame with the expected
columns will work in a sonification pipeline.
Users can define their own custom \code{rollup} functions,
including those that duplicate rows to generate chords or
parallel voices.
}
