% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-midi.R
\name{write_midi}
\alias{write_midi}
\title{Write a data frame of notes to a Standard MIDI file}
\usage{
write_midi(notes, filename = "test.mid", opt = list())
}
\arguments{
\item{notes}{A data frame containing columns
\code{tick_on}, \code{tick_off}, \code{pitch}, and \code{velocity}.
\code{tick_on} and \code{tick_off} should be integers.
\code{pitch} and \code{velocity} should be integers.
\code{channel} and \code{group} should be factors; other columns should be numeric.}

\item{filename}{Name of the output file.}

\item{opt}{A list of options.}
}
\value{
The path to the output file is invisibly returned.
}
\description{
Write a data frame of notes to a Standard MIDI file
}
